Split GetConfig and IConfig
